backend/gmail_utils.py
import base64
import os
from google.oauth2.credentials import Credentials
from googleapiclient.discovery import build

def connect_to_gmail():
    """
    Connect to Gmail API using stored credentials.
    This is a placeholder for now.
    """
    creds_path = "gmail_token.json"
    if not os.path.exists(creds_path):
        return None

    creds = Credentials.from_authorized_user_file(creds_path, ["https://mail.google.com/"])
    service = build("gmail", "v1", credentials=creds)
    return service

def fetch_latest_message():
    """
    Example: Fetch the latest Gmail message.
    """
    service = connect_to_gmail()
    if not service:
        return {"error": "No Gmail credentials found."}

    results = service.users().messages().list(userId="me", maxResults=1).execute()
    message_id = results["messages"][0]["id"]
    message = service.users().messages().get(userId="me", id=message_id, format="full").execute()

    snippet = message.get("snippet", "")
    return {"latest_email_snippet": snippet}
